// frontend/src/shared/server/types.ts
export enum ModbusErrorStates {
  noerror,
  timeout,
  crc,
  other,
  illegalfunctioncode,
  illegaladdress,
  initialConnect,
}

export interface ImodbusAddress {
  address: number
  registerType: ModbusRegisterType
  write?: number[]
  length?: number
}
export enum ModbusTasks {
  deviceDetection = 0,
  splitted = 1,
  tcpBridge = 2,
  poll = 3,
  specification = 4,
  entity = 5,
  writeEntity = 6,
  initialConnect = 7,
}
export interface ImodbusErrorsForSlave {
  task: ModbusTasks
  date: number
  address: ImodbusAddress
  state: ModbusErrorStates
}
export interface ImodbusStatusForSlave {
  requestCount: number[]
  errors: ImodbusErrorsForSlave[]
  queueLength: number
}